Fix some compiler warnings
Warnings about using the wrong type, wrong types being passed to printf-like functions, etc.

Remove useless casts

+++ sftp-client.c
@@ -2103,7 +2103,7 @@ do_upload(struct sftp_conn *conn, const char *local_path,
 	if ((msg = sshbuf_new()) == NULL)
 		fatal_f("sshbuf_new failed");
 	for (;;) {
-		int len;
+		ssize_t len;
 
 		/*
 		 * Can't use atomicio here because it returns 0 on EOF,
@@ -2122,14 +2122,14 @@ do_upload(struct sftp_conn *conn, const char *local_path,
 			fatal("read local \"%s\": %s",
 			    local_path, strerror(errno));
 		} else if (len != 0) {
-			ack = request_enqueue(&acks, ++id, len, offset);
+			ack = request_enqueue(&acks, ++id, (size_t)len, offset);
 			sshbuf_reset(msg);
 			if ((r = sshbuf_put_u8(msg, SSH2_FXP_WRITE)) != 0 ||
 			    (r = sshbuf_put_u32(msg, ack->id)) != 0 ||
 			    (r = sshbuf_put_string(msg, handle,
 			    handle_len)) != 0 ||
 			    (r = sshbuf_put_u64(msg, offset)) != 0 ||
-			    (r = sshbuf_put_string(msg, data, len)) != 0)
+			    (r = sshbuf_put_string(msg, data, (size_t)len)) != 0)
 				fatal_fr(r, "compose");
 			send_msg(conn, msg);
 			debug3("Sent message SSH2_FXP_WRITE I:%u O:%llu S:%u",
